requirements.txt
<<<<<<< HEAD
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
cssselect==1.3.0
cssutils==2.11.1
et_xmlfile==2.0.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
httplib2==0.30.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.1
MarkupSafe==3.0.2
more-itertools==10.7.0
numpy==2.3.2
oauth2client==4.1.3
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.2
premailer==3.10.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
soupsieve==2.8
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
yagmail==0.15.293
=======
streamlit>=1.20
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
>>>>>>> 368c19b (Initial commit for Streamlit GSC app)
